# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Furniture ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"

    post:
      operationId: createRole
      tags: [roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBody"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"

    put:
      operationId: updateRole
      tags: [roles]
      summary: Update role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"

    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # CATEGORIES
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierListResponse"

    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # MANUFACTURERS
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List purchase orders
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [supplier, manufacturer]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, in_production, shipped, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderListResponse"

    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderBody"
      responses:
        "201":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      summary: Get purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

    put:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      summary: Update purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      summary: Delete purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /purchase-orders/{id}/status:
    patch:
      operationId: updatePurchaseOrderStatus
      tags: [purchase-orders]
      summary: Update purchase order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # DASHBOARD
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PermissionSet"
        createdAt:
          type: string
          format: date-time
      required: [id, name, permissions, createdAt]
    PermissionSet:
      type: object
      properties:
        view:
          type: boolean
        add:
          type: boolean
        edit:
          type: boolean
        delete:
          type: boolean
      required: [view, add, edit, delete]
    CreateRoleBody:
      type: object
      properties:
        name:
          type: string
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PermissionSet"
      required: [name, permissions]

    # CATEGORIES
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: integer
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        price:
          type: number
        gstPercent:
          type: number
        stockQty:
          type: integer
        lowStockThreshold:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, sku, price, gstPercent, stockQty, lowStockThreshold, createdAt]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        gstNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, isActive, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        gstNumber:
          type: string
          nullable: true
      required: [name]
    SupplierListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Supplier"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # MANUFACTURERS
    Manufacturer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        specialization:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, isActive, createdAt]
...
    PurchaseOrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [id, productId, quantity, unitPrice]
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        poNumber:
          type: string
        type:
          type: string
          enum: [supplier, manufacturer]
        supplierId:
          type: integer
          nullable: true
        supplier:
          $ref: "#/components/schemas/Supplier"
          nullable: true
        manufacturerId:
          type: integer
          nullable: true
        manufacturer:
          $ref: "#/components/schemas/Manufacturer"
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, in_production, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        totalAmount:
          type: number
        expectedDelivery:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, poNumber, type, status, items, totalAmount, createdAt]
    CreatePurchaseOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
    CreatePurchaseOrderBody:
      type: object
      properties:
        type:
          type: string
          enum: [supplier, manufacturer]
        supplierId:
          type: integer
          nullable: true
        manufacturerId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePurchaseOrderItemBody"
        expectedDelivery:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
      required: [type, items]
    UpdatePurchaseOrderBody:
      type: object
      properties:
        expectedDelivery:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
    UpdatePurchaseOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, in_production, shipped, delivered, cancelled]
      required: [status]
    PurchaseOrderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrder"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # DASHBOARD
...